Backend/Capstone/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2